+++ test/integration/default/kitchen_config_spec.rb
@@ -1,0 +1,8 @@
+describe file('C:/Users/Vagrant/.kitchen/config.yml') do
+  if ENV['http_proxy']
+    it { should exist }
+    its('content') { should match(/<%/) }
+  else
+    it { should_not exist }
+  end
+end